demo captcha canvas element
When inspecting the demo page for the js-captcha, I came across this canvas element because it was in the source code when I used web inspector on my Mac. Maybe the only possible reason I saw the math problem was due to the fact that GitHub hosts these websites for us. If you have an Apple computer try adding the demo code to a (not hosted) .html file.